+++ angular/app/landing/landing.js
@@ -30,14 +30,6 @@
 			}
 		}, 2000);
 
-		$scope.alertDialog = function(){
-			DialogService.alert('This is an alert title', 'You can specify some description text in here.');
-		};
-
-		$scope.customDialog = function(){
-			DialogService.fromTemplate('add_users');
-		};
-
 	});
 
 })();